requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
google-cloud-aiplatform>=1.90.0
google-cloud-firestore>=2.20.0
google-cloud-translate>=3.18.0
googlemaps>=4.10.0
python-dotenv>=1.1.0
pydantic>=2.11.0
slowapi>=0.1.9
pytest>=8.3.0
httpx>=0.28.0
